feat(event-display): use class for color options in UI
